Refactor and style
